backend/requirements.txt
# Ultra-Lean Requirements for $14.99/month SaaS with 70% margins
# Only essential dependencies for premium-feeling asthma coach

# FastAPI Core
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0

# Database & Auth
supabase>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0

# Data & ML Core (Ultra-lightweight)
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0

# Single ML Model
xgboost>=2.0.0,<3.0.0

# Explainability (for scientific feel)
shap>=0.42.0,<1.0.0

# API Clients
httpx>=0.25.0,<1.0.0
requests>=2.31.0,<3.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
tqdm>=4.65.0,<5.0.0

# Security (CRITICAL - DO NOT REMOVE)
bleach>=6.0.0,<7.0.0  # HTML sanitization (XSS prevention)
python-jose[cryptography]>=3.3.0,<4.0.0  # JWT tokens (already included above)
passlib[bcrypt]>=1.7.4,<2.0.0  # Password hashing (already included above)

# Explicitly REMOVED to save costs:
# tensorflow, keras (heavy deps)
# transformers, torch (expensive API costs)
# prophet, darts, lightgbm, catboost (redundant)
# flower (distributed compute)
# jupyter (development only)
# gymnasium, ray (RL overkill)
# AutoML packages (sklearn covers it)